pyproject.toml
[build-system]
requires = ['setuptools>=45', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = 'city'
version = '0.0.0'
description = 'A generational simulator for LLM societies'
readme = 'README.md'
requires-python = '>=3.9'
dependencies = [
    'openai>=2.16.0',
    'anthropic>=0.77.0',
    'python-dotenv>=0.9.0',
    'pytest>=9.0.0',
    'pytest-cov>=4.0.0',
    'sqlalchemy>=2.0.0',
]

[project.scripts]
city = 'city.__main__:main'

[tool.pytest.ini_options]
testpaths = ['tests']
pythonpath = ['.']
